lots of tree cleanup. - introduce scoped_ptr in some places. internal usage only. - refactor TempFile to be ref counted intrusive_ptr - header cleanup. engine.h, dom.h - rename util to base, associated munging - prefix DISALLOW_EVIL_CONSTRUCTORS with LIBKML_ to avoid macro clashing - a few more engine examples
note to self: commit more often. :-)

+++ examples/helloworld/hellofolder.cc
@@ -63,14 +63,14 @@ int main(int argc, char** argv) {
   // Create a Folder, fill it with Placemarks.
   cout << "Creating " << kHowManyPoints << " Placemarks:" << endl;
   FolderPtr folder = factory->CreateFolder();
-  for (int i = 0 ; i < kHowManyPoints ; ++i) {
+  for (size_t i = 0 ; i < kHowManyPoints ; ++i) {
     folder->add_feature(CreatePlacemark(factory, "hi", 1.1, 2.2));
   }
 
   // Read back all the Placemarks in the Folder.
   cout << "Iterating " << kHowManyPoints << " Placemarks:" << endl;
   assert(kHowManyPoints == folder->get_feature_array_size());
-  for (int i = 0; i < folder->get_feature_array_size(); ++i) {
+  for (size_t i = 0; i < folder->get_feature_array_size(); ++i) {
     assert("hi" == folder->get_feature_array_at(i)->get_name());
     assert(kmldom::Type_Placemark == folder->get_feature_array_at(i)->Type());
   }
